refactor: clean up code formatting and remove unused includes

+++ src/core/pipe/pipe_manager.cpp
@@ -159,7 +159,7 @@ void PipeManager::handleClient(HANDLE hPipe)
                 // Send acknowledgment back to client
                 std::string response = "OK: " + feature + ":" + (enabled ? "enabled" : "disabled");
                 DWORD bytesWritten;
-                WriteFile(hPipe, response.c_str(), response.length(), &bytesWritten, nullptr);
+                WriteFile(hPipe, response.c_str(), static_cast<DWORD>(response.length()), &bytesWritten, nullptr);
                 FlushFileBuffers(hPipe);
             }
             else
@@ -167,7 +167,7 @@ void PipeManager::handleClient(HANDLE hPipe)
                 LOG_ERROR("Unknown command: '{}'", cmd.c_str());
                 std::string response = "ERROR: Unknown command";
                 DWORD bytesWritten;
-                WriteFile(hPipe, response.c_str(), response.length(), &bytesWritten, nullptr);
+                WriteFile(hPipe, response.c_str(), static_cast<DWORD>(response.length()), &bytesWritten, nullptr);
                 FlushFileBuffers(hPipe);
             }
         }
@@ -176,7 +176,7 @@ void PipeManager::handleClient(HANDLE hPipe)
             LOG_ERROR("Invalid message format: '{}'", message.c_str());
             std::string response = "ERROR: Invalid format";
             DWORD bytesWritten;
-            WriteFile(hPipe, response.c_str(), response.length(), &bytesWritten, nullptr);
+            WriteFile(hPipe, response.c_str(), static_cast<DWORD>(response.length()), &bytesWritten, nullptr);
             FlushFileBuffers(hPipe);
         }
     }
